+++ playground/site/templates/exception.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * This template is used to trigger a Kirby exception.
+ * It is used to test the exception notifications.
+ */
+
+throw new Exception('This is a test exception');